apps/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asteroid-filterbanks==0.4.0
attrs==25.4.0
audioop-lts==0.2.2
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
einops==0.8.1
email-validator==2.3.0
fastapi==0.121.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
jose==1.0.0
julius==0.2.7
jwt==1.4.0
kiwisolver==1.4.9
lightning==2.5.6
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numba==0.63.1
numpy==2.3.4
openai-whisper==20250625
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.6.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytorch-lightning==2.5.6
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
SQLAlchemy==2.0.44
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchcodec==0.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
zipp==3.23.0
